+++ packages/electron-esbuild/src/builder/base.ts
@@ -5,14 +5,14 @@
  */
 
 import type { Builder } from '../builder'
-import type { ConfigItem } from '../config/config'
+import type { Item } from '../config/config'
 
 export abstract class BaseBuilder<T> implements Builder {
   readonly env: string
 
   abstract readonly hasInitialBuild: boolean
 
-  protected constructor(protected readonly _config: ConfigItem<T>) {
+  protected constructor(protected readonly _config: Item<T>) {
     this.env = this._config.isMain
       ? 'Main'
       : this._config.isRenderer