+++ packages/electron-esbuild/src/config/configurators/vite.ts
@@ -7,14 +7,14 @@
 import nodeModule from 'module'
 import { InlineConfig } from 'vite'
 
-import type { ItemConfig } from '../config'
+import type { EnvConfig } from '../config'
 import { TypeConfig } from '../enums'
 import type { Configurator } from './base'
 
-export class ViteConfigurator implements Configurator<TypeConfig.Vite> {
-  public readonly type = TypeConfig.Vite
+export class ViteConfigurator implements Configurator<TypeConfig.vite> {
+  public readonly type = TypeConfig.vite
 
-  constructor(public readonly config: ItemConfig) {}
+  constructor(public readonly config: EnvConfig) {}
 
   toBuilderConfig(partial: Partial<InlineConfig>): InlineConfig {
     let external = partial?.build?.rollupOptions?.external