+++ packages/electron-esbuild/src/config/configurators/webpack.ts
@@ -8,14 +8,14 @@ import deepMerge from 'deepmerge'
 import path from 'path'
 import { Configuration } from 'webpack'
 
-import type { ItemConfig } from '../config'
+import type { EnvConfig } from '../config'
 import { Target, TypeConfig } from '../enums'
 import type { Configurator } from './base'
 
-export class WebpackConfigurator implements Configurator<TypeConfig.Webpack> {
-  public readonly type = TypeConfig.Webpack
+export class WebpackConfigurator implements Configurator<TypeConfig.webpack> {
+  public readonly type = TypeConfig.webpack
 
-  constructor(public readonly config: ItemConfig) {}
+  constructor(public readonly config: EnvConfig) {}
 
   toBuilderConfig(
     partial: Partial<Configuration>,
@@ -26,7 +26,7 @@ export class WebpackConfigurator implements Configurator<TypeConfig.Webpack> {
       partial,
       {
         output: {
-          filename: target === Target.Main ? 'main.js' : 'index.js',
+          filename: target === Target.main ? 'main.js' : 'index.js',
           path: path.resolve(process.cwd(), this.config.output),
         },
       },