+++ packages/electron-esbuild/src/config/validation.ts
@@ -7,21 +7,22 @@
 import Joi from 'joi'
 
 import { Logger } from '../console'
+import { EnvConfig } from './config'
 import { TypeConfig } from './enums'
-import { YamlSkeleton } from './yaml'
+import { Yaml, YamlSkeleton } from './yaml'
 
 const _logger = new Logger('Config/Validation')
 
 const _schema = Joi.object<YamlSkeleton>({
   mainConfig: Joi.object({
-    type: Joi.string().valid(TypeConfig.Esbuild).required(),
+    type: Joi.string().valid(TypeConfig.esbuild).required(),
     path: Joi.string().required(),
     src: Joi.string().required(),
     output: Joi.string().required(),
   }).required(),
   rendererConfig: Joi.object({
     type: Joi.string()
-      .valid(TypeConfig.Esbuild, TypeConfig.Webpack, TypeConfig.Vite)
+      .valid(TypeConfig.esbuild, TypeConfig.webpack, TypeConfig.vite)
       .required(),
     path: Joi.string().required(),
     src: Joi.string().required(),
@@ -30,15 +31,33 @@ const _schema = Joi.object<YamlSkeleton>({
   }).optional(),
 })
 
-export function validate(config: unknown): true | never {
-  const result = _schema.validate(config)
-
-  if (result.error) {
-    return _logger.end(
-      'Configuration file contains errors',
-      result.error.details.map((item) => item.message).join('; '),
-    )
+export class ConfigFile {
+  constructor(public readonly config: YamlSkeleton) {
+    if (this.config.rendererConfig === undefined) {
+      this.config.rendererConfig = null
+    }
   }
 
-  return true
+  ensureValid(): true | never {
+    const result = _schema.validate(this.config)
+
+    if (result.error) {
+      return _logger.end(
+        'Configuration file contains errors',
+        result.error.details.map((item) => item.message).join('; '),
+      )
+    }
+
+    return true
+  }
+
+  toYaml(): Yaml {
+    return new Yaml({
+      main: EnvConfig.fromYaml(this.config.mainConfig),
+      renderer:
+        this.config.rendererConfig !== null
+          ? EnvConfig.fromYaml(this.config.rendererConfig)
+          : null,
+    })
+  }
 }