+++ packages/electron-esbuild/src/index.ts
@@ -11,7 +11,7 @@ import meow from 'meow'
 import { Cli, CliFlags } from './cli'
 import { commands } from './commands'
 
-const cli = meow<CliFlags>(
+const _cli = meow<CliFlags>(
   `Usage
   $ electron-esbuild [command]
 
@@ -37,19 +37,19 @@ Examples
 
 type Commands = keyof typeof commands
 
-const command = cli.input[0] ?? null
-const availableCommands = ['dev', 'build']
+const _command = _cli.input[0] ?? null
+const _availableCommands = ['dev', 'build']
 
 function isValidAction(command: string): command is Commands {
-  if (command === null || !availableCommands.includes(command)) {
-    cli.showHelp(0)
+  if (command === null || !_availableCommands.includes(command)) {
+    _cli.showHelp(0)
   }
 
   return true
 }
 
-if (isValidAction(command)) {
-  const action: Cli = new commands[command](cli)
+if (isValidAction(_command)) {
+  const action: Cli = new commands[_command](_cli)
 
   action.init().then(() => {
     // process.exit(0)